git:chore - move CommitAuthor declaration to git pkg
